PHASE 1 finalized

+++ src/jsmith/jbt/com/DailyCouponExpirationTask.java
@@ -5,6 +5,7 @@ package jsmith.jbt.com;
 
 import java.sql.Date;
 import java.util.GregorianCalendar;
+import java.util.concurrent.TimeUnit;
 
 import jsmith.jbt.com.DAO.CompanyCouponDBDAO;
 import jsmith.jbt.com.DAO.CouponDBDAO;
@@ -24,23 +25,28 @@ public class DailyCouponExpirationTask implements Runnable {
 	private final CompanyCouponDBDAO compcouponDBDAO;
 	private final CouponDBDAO couponDBDAO;
 	private final ConnectionPool cPool;
+	private final TimeUnit tuPeriod;
+	private final  long periodValue;
+	private Date lastRun=null;
 	/* (non-Javadoc)
 	 * @see java.lang.Runnable#run()
 	 */
 	@Override
 	public void run() {
 		while(!quit){
-			try {
-				Date today=new Date(GregorianCalendar.getInstance().getTimeInMillis());
-				//TODO check the Time - is it time to run the scheduled cleaning
-				for(Coupon outdatedCoupon: couponDBDAO.realAllByEndDate(today)) {
-					custcouponDBDAO.deleteAllCoupons(outdatedCoupon.getID()); //Delete from all customers
-					compcouponDBDAO.deleteAllCoupons(outdatedCoupon.getID()); //Delete from company
-					couponDBDAO.delete(outdatedCoupon);						  //Delete the entity record
-				}
-				Thread.sleep(5000); //TODO remove the sleep
-			} catch (InterruptedException | CouponSystemException e) {
-				System.out.println("DailyExpirationRoutine interrupted by force, quit flag="+this.quit);
+			Date today=new Date(GregorianCalendar.getInstance().getTimeInMillis());
+			if(lastRun==null || CouponDbHelper.computeDetailedDateDiff(today, lastRun).get(tuPeriod)>periodValue) {
+				try {
+					for(Coupon outdatedCoupon: couponDBDAO.realAllByEndDate(today)) {
+						custcouponDBDAO.deleteAllCoupons(outdatedCoupon.getID()); //Delete from all customers
+						compcouponDBDAO.deleteAllCoupons(outdatedCoupon.getID()); //Delete from company
+						couponDBDAO.delete(outdatedCoupon);						  //Delete the entity record
+						lastRun=today;
+						Thread.sleep(1000);
+					}
+				} catch (InterruptedException | CouponSystemException e) {
+					System.out.println("DailyExpirationRoutine interrupted by force, quit flag="+this.quit);
+				}				
 			}
 		}
 		
@@ -49,13 +55,15 @@ public class DailyCouponExpirationTask implements Runnable {
 	 * @throws CouponSystemException 
 	 * 
 	 */
-	public DailyCouponExpirationTask() throws CouponSystemException {
+	public DailyCouponExpirationTask(TimeUnit tuPeriod,long periodValue) throws CouponSystemException {
 		super();
 		quit=false;
 		this.cPool=ConnectionPool.getInstance(ConnectionPool.defDriverName, ConnectionPool.defDbUrl);
 		this.custcouponDBDAO=new CustomerCouponDBDAO(cPool);
 		this.compcouponDBDAO=new CompanyCouponDBDAO(cPool);
 		this.couponDBDAO=new CouponDBDAO(cPool);
+		this.tuPeriod=tuPeriod;
+		this.periodValue=periodValue;
 	}
